Remove unneeded CoreApi
